feat/styles estilos movidos de provider a utils funcionando
